% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{calibration_plot}
\alias{calibration_plot}
\title{Create Analysis Plots}
\usage{
calibration_plot(rf.df.1, target, risk_col, vars_list)
}
\arguments{
\item{rf.df.1}{the dataframe}

\item{target}{the column name of the target variable}

\item{risk_col}{the column with the probabilty of having the outcome of interest}

\item{vars_list}{the variables used in training the random forest}
}
\value{
shows the calibration and rpart plots for the model
}
\description{
\code{analysis_plots} creates plots of both the calibration of the random forest and the rpart summary tree
}
