% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{plot_time_varying_auc}
\alias{plot_time_varying_auc}
\title{Plot time varying AUC}
\usage{
plot_time_varying_auc(sca1.df = data, target, patient_count_col, time_col)
}
\arguments{
\item{sca1.df}{the data}

\item{target}{String variable containing the name of the target}

\item{patient_count_col}{String variable containing the name of the column couting the patient number}

\item{time_col}{String variable containing the name of the column holding the CPIU number}
}
\value{
shows a time-varying auc plot
}
\description{
\code{rf.auc} calculates the time varying AUC values and returns them in a dataframe
}
