% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_functions_p2.R
\name{rpart_summary}
\alias{rpart_summary}
\title{Create Analysis Plots}
\usage{
rpart_summary(rf.df.1, risk_col, vars_list)
}
\arguments{
\item{rf.df.1}{the dataframe}

\item{risk_col}{the column with the probabilty of having the outcome of interest}

\item{vars_list}{the variables used in training the random forest}
}
\value{
shows the rpart plot for the model
}
\description{
\code{analysis_plots} creates plots of both the calibration of the random forest and the rpart summary tree
}
